fix(changelog): remove changes related to unreleased commits (#68)
There was a bug with v1.0.1-next.1 tag pointing at a wrong commit hash, which resulted in the last two
releases to NPM fail. This bug was caused by an attempt to update a semantic-release-bot commit and force-pushing it to 'next' branch. The commit hash changed and so semantic release had no reference to the commit tagged by v1.0.1-next.1.

The solution was to force-update the referenced commit hash in v1.0.1-next.1 tag.